Fix the multiple load issue for composites
The composite loading would crash when several composites would be loaded
one after the other. This was because composite config files where loaded
partially but were considered loaded entirely. In order to fix this
problem and make things simpler, we removed the composite config mechanism
entirely, so that the composites are reloaded everytime. That allows both
config changing on the fly, but also more resilience for multiple sensor
cases, like one sensor is loaded after another, and the composites wouldn't
get updated.

Signed-off-by: Martin Raspaud <martin.raspaud@smhi.se>