Fix DatasetDict's setitem to allow empty md in value.
Sometimes a dataset/projectable doesn't have any info attached to it, eg
because the dataset is synthetic. In these cases, setitem would crash.
This is now fixed, and if a string is provided as a key in setitem it is
used as a name if no better name is already there.

Signed-off-by: Martin Raspaud <martin.raspaud@smhi.se>