Allow resample to receive strings as area identifiers
In resample, the interactive user would most likely use pre-defined areas
from a custom area file. In this case, it's much easier to refer to the
area by name, than to get the area definition object from the file. This
patch allows the `resample` projectable method to work with string ids
also.

Signed-off-by: Martin Raspaud <martin.raspaud@smhi.se>