Change LI L2 to xarray
Change the LIFileHandler.get_dataset method to return an
xarray DataArray rather than a masked array.  The module now imports
h5py rather than h5netcdf.  Masking is now implicit because with xarray
DataArray, we are masking with NaN.  Information from the key passed
to get_dataset is no longer stored in the returned object because I do
not yet know what should be stored where and how.

This still needs a test.