Add links to source from API documentation
From the API documentation, add links the source for each class, method,
and function.  In a perfect world, each of those would have perfect
documentation such that a developer will not need to refer to the
source.  I believe that some parts of satpy are not quite there yet, and
therefore such links to the source could be useful in the interim.