Correct ranges, scaling, and units
viirs_gran was originally a cut-and-paste of the viirs_l1b reader. Since the file formats are similar, this was a good first approach. However, the ranges, scale factors, and units are stored differently in gran files than in L1b. This caused viirs_gran to fail on variables where the units were percent or Kelvin (e.g. ice concentration).

Resolved by correcting formats to be consistent with those in the gran data, and by removing unnecessary unit checks.