Add test to ensure FCI reader logs no warnings
Add a test to ensure that using the FCI reader does not lead to any log
messages at warning level or above.

Although I expect this test to fail as I haven't fixed #1171 yet, it is
currently failing for the wrong reason as despite running with pytest,
it doesn't seem to understand using the caplog fixture.