FCI: read pixel quality
Add pixel quality fields for FCI reading.  This currently relies on a
small hack rewriting the ancillary_variables fields, because the data
variable pixel_quality has the same name in different groups, and Satpy
doesn't currently have a good way of keeping track of what group this
belongs to.

Fixes 1171.