Add per-frame decoration to MultiScene
In MultiScene.save_animation, add the ability to add decorations,
enhancements, and overlays.  As a special case, decoration text will be
substituted allowing it to vary per frame.  This enables the user to add
the time to every frame.

I took the unit test out of the `unittest.TestCase` class because I
could otherwise not get the call assertions on the `get_enhanced_image`
mock to wark.