Replace assert_called_with with direct args check
In the mock for testing decorating a multiscene, replace the assertion
that it's called with a particular dataset to more direct argument
checking.  Any assert_calls with a numpy array is unreliable, because
assert_called_with will use an equality operator which may fail, in this
case depending on the unpredictable order of calls.  See
https://stackoverflow.com/q/56644729/974555