Added test for multiple readers in multiscene
Added a test for using multiple readers in a MultiScene.  Note that
the corresponding functionality is not yet implemented and actually
not even yet properly defined (see #1268).  To implement this in
`MultiScene.from_files` requires updating `satpy.readers.group_files`.

This commit does change the implementation of the keyword-only argument in
`satpy.readers.group_files` by using Pythons own keyword-only argument
functionality (available since Python 3.0).  This has the effect that
failing to pass the `reader` to that function now raises a TypeError
rather than previously a ValueError.