Update sensor_names method and values The sensor name is not defined in the file.  Make an attempt to get the attribute from the file, but provide the information if not retrieved from the netcdf file.
Rename the dimensions to x,y to keep consistent with satpy.

Update the sensor list in both the yaml file and add to the
code so that a failure to read the sensor from the netCDF
file cascades to an actual sensor list.
prefer these variables for the dimensions (the netCDF reader).