Make FSFile open in bytes mode
When FSFile opens a file, it should open it in bytes mode, not in text
mode.  Most of our data are binary.  If there are any readers that need
files opened in text mode we need to find a way to deal with that.

The unit test is not entirely desirable, because all it does is create
an empty file and see that the expected exception is raised.  I did not
succeed in suppressing the warning issued by pytest that another
exception is raised on destroying the object.  The catch_warnings
context manager couldn't help me here, because this destruction happens
after the unit test has already completed (I think).