Improve MiRS reader handling of missing metadata (#1671)
* Add reader_kwarg to omit limb correction on ATMS sensors.

* Fix docstring in mirs reader class

* add check for calling limb correction only when sensor is atms

* Fix docstring

* Add a kwarg test for limb_correction.

* Remove unused variables and add a noaa-20 test
Remove platform name from parameterize in last test for kwargs because
they are not used in that test.  Add a test for the NOAA-20 platform
when the limb_correction assertions are tested and platform name is
tested.

* Fold reader_kwarg test into basic_load

* Simply if/then statement, split parameterize for readability.

* Simplify assertion if/then statement for limb_correction.

* Remove extra line getting the name of the sensor

* Use valid range when present
Fixes a bug in which valid range was ignored.

* Check to confirm that valid range is no longer in attributes.

* Add a test to check valid range was applied correctly.
Change get_valid_range=>apply_valid_range and check for case of valid_range not None

* valid_range is inclusive so include both min/max in acceptable values
only read fake_data one time, not every time check_valid_range is called.

* Add attributes when missing/apply attributes from both file and yaml including _FillValue scale.
- Fill value needs to be scaled in these files.
- Try to combine yaml and file attributes with file attributes taking precedence over yaml.
- Make sure units is saved in data array attributes

* Fix _FillValue so that it is read as an integer
Fix name of units

* Update reading of yaml
- Need to make sure that yaml metadata when present is used and supplements
  attributes from data_arr.
- Apply scaling to valid_range and fill_value
- move location of apply_attributes
- change name of get_metadata

* Test units for TEST_VARS

* Add descriptions for some of the variables in the yaml.

* BUG FIXES: apply attributes before limb correction and fix typos in yaml
- apply the attributes before the limb correction so that scaling, valid_range, fill are applied correctly
- the yaml had some copy and pasted errors that would make it useless and unreadable
- remove BT from yaml and handle attributes in reader
- use the yaml for lat/lon attributes from every file

* Remove the change to the file_patterns, the extra file_type is not needed for mirs_gpm
Added because I did not consider that the limb correction is applied based on sensor which
is assigned by platform.  The mirs_gpm addition was not needed so I removed it.

* Add BT to yaml and use in creation of ds_info for data_id

* commit the yaml mentioned in previous commit

* Take file_key out of BT dataset so that it does not get carried through to btemp dataset information

* Don't add more in yaml than necessary
If the information is in the metadata of the file, use that, otherwise
define in the yaml.  Fix getting keys from ds_info in apply_attributes
function so it does not needlessly break data that has no units.

* Make sure btemp information is only initializing with yaml when necessary.

* Simplify the reading of coefficients so reading old/versus new coefficients is consistent
Old coefficients have a blank line at end, new coefficients do not.  Previously, n_chn and
n_fov were fixed in the coefficient reading.  So, return to that so it is easier to swap
new and old coefficients if necessary for testing.

* Take out check for n_chn and n_fov since they are fixed now.

* Simplify logic which repeatedly checked if file_type matched yaml datasets.