Add functionality to check on empty groups
When passing multiple readers to group_files, allow the user to control
the behaviour if one or more groups have no files for one or more
readers.  The default, "pass", is the status quo, and will result in
groups where one or more readers have zero files.  The other options are
"raise", which will raise an exception if this happens, and "skip",
which will remove such groups from the result completely.