Changed implementation of _get_sorted_file_groups
Simplified the implementation of _get_sorted_file_groups to yield the
retained groups rather than remove the removed groups from a list
in-situ.  Make sure we raise a ValueError if an invalid value is passed
for the missing keyword argument.  Test more specifically for the error
message coming with the exception raised.