Fix ABI L1B reader not returning expected attrs.
As part of Pull Request #1692, we added an early return to abi_l1b.py's
get_dataset method when the requested calibration is "counts".

The problem is that the early return results in a number of missing
attributes in the object returned from get_dataset.

I also took the liberty of updating the unit tests to check for some of
the missing attributes, as well as including a test for invalid
calibration.