Start working on ninjogeotiff
Very first start working on a writer that writes a regular GeoTIFF file
but with NinJo tags added according to the format agreed with the NinJo
team.  Although this will retire pyninjotiff and ninjotiff, it will
still need its own writer because some of the tags will need to be
calculated on the fly as a function of the content of the data to be
written.  So far this contains only a broken and breaking unit test that
isn't even testing the right thing yet, so even if nothing written so
far might even survive, at least it's a start ☺.