Extend unit tests for parallax correction.
Add unit tests for the low-level parallax_correct function, according to
an interface I believe it should have, but that it doesn't have.  Most
of the unit tests are failing.

The higher level class is not unit tested yet because I don't understand
what it's supposed to do exactly.  Probably much of it needs to be
deleted.