Migrate all PalettizeCompositor composites
Migrate all PalettizeCompositor composites to use SingleBandCompositor
and palettize instead, using the new functionality where the palette is
read from the dataset contained with the data.  This mostly
(exclusively?) affects NWCSAF imagery.  It is now possible relatively
easily to produce a p-mode image::

    from satpy import Scene
    sc = Scene(
	filenames=["/media/nas/x21308/scratch/NWCSAF/NWCSAF-NinjoTIFF-2/S_NWC_CT_MSG4_MSG-N-VISIR_20210930T100000Z.nc"],
	reader=["nwcsaf-geo"])
    sc.load(["cloudtype"])
    sc.save_datasets(keep_palette=True)

Also improve documentation for this functionality and note the deprecation of the old ways.