Do not set min/max value for palettize enhancements
Do not set min/max value for palettize enhancements when the palette is
obtained from the source file!  The source file contains a valid_range
which is already used in the construction of the palette.  Setting
min/max value will destroy the match and result in colours that are
doubleplusungood.