Use shared rather than first wishlist
When blending a multiscene, build the new wishlist from the wishlists
shared between the scenes, analogous to the common datasets.  If we just
copy the wishlist from the first scene, generating possible composites
will fail if there are datasets in the first scene that are not in the
multiscene because they don't have the same datasets.