Prepare yaml reader for preloading filehandlers
Prepare GeoSegmentYAMLReader for preloading filehandlers for files
corresponding to segments that don't exist yet.  Early draft
implementation that appears to work with limitations.  Implementation in
GEOSegmentYAMLReader still needs tweaking, tests need to be improved,
and the corresponding file handler (for now just FCI) needs to be able
to handle it.