Progress on Preloadable
Add a Preloadable class to netcdf_utils.  This so far implements
pre-loading filehandlers for to-be-expected files if a single one
already exists, taking a defined set of data variables and their
attributes from the first segment.  Still to be implemented is to take
other information from other repeat cycles, by on-disk caching.