Move composite generation back to BaseScene
The composite generation differs only in name from the equivalent for a
VectorScene.  For ease of implementation, move this back to the
BaseScene.

Make the GeoDataFrame to be tested with not empty.  It looks like
concatenating an empty GeoDataFrame with pandas results in a DataFrame,
but if it's non-empty than the type GeoDataFrame is retained.

The first test passes now.